requirements.txt
astroid==3.3.8
Click==8.1.8
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
lazy-object-proxy==1.10.0
MarkupSafe==3.0.2
mccabe==0.7.0
pandas==2.2.3
PyGithub==2.5
PyShExC==0.9.1
pytest~=8.3.4
python-dateutil~=2.9.0
PyYAML==6.0.2
recommonmark==0.7.1
six==1.17.0
Sphinx==7.4.7
sphinx-rtd-theme==3.0.2
sphinx-markdown-tables==0.0.17
stsci-rtd-theme==1.0.1
rdflib==7.1.3
requests==2.32.3
setuptools~=75.8.0
tqdm==4.67.1
ontospy==2.1.1
unidiff==0.7.5
urllib3==2.3.0
Werkzeug==3.1.3
wikidataintegrator==0.9.30
wrapt==1.17.2